% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ideanet_viz}
\alias{ideanet_viz}
\title{Launch Shiny App}
\usage{
ideanet_viz(name = "ideanet_viz", ...)
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch Shiny App
}
