% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ideanetViz}
\alias{ideanetViz}
\title{Launch Shiny App}
\usage{
ideanetViz(name = "ideanetViz", ...)
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch Shiny App
}
